msm: camera: ov13850: Don't read OTP data during probe
Willi Ye noticed the broken actuator behaviour on a5 and memul is
influenced by the compiler version used, in that the kernel built with
GCC 4.8 reads completely bogus OTP data:

GCC 4.7:
<6>[    0.585531] c0      1 [CAM] ov13850_read_fuseid: OTP valid layer = 0
<6>[    0.585812] c0      1 [CAM] ov13850_read_fuseid: OTP Module vendor = 0x3
<6>[    0.585823] c0      1 [CAM] ov13850_read_fuseid: OTP LENS = 0x61
<6>[    0.585833] c0      1 [CAM] ov13850_read_fuseid: OTP Sensor Version = 0x1
<6>[    0.585843] c0      1 [CAM] ov13850_read_fuseid: OTP Driver IC Vendor & Version = 0x31
<6>[    0.585854] c0      1 [CAM] ov13850_read_fuseid: OTP Actuator vender ID & Version = 0x47

GCC 4.8:
<6>[    0.581601] c0      1 [CAM] ov13850_read_fuseid: OTP valid layer = 2
<6>[    0.581858] c0      1 [CAM] ov13850_read_fuseid: OTP Module vendor = 0xd4
<6>[    0.581868] c0      1 [CAM] ov13850_read_fuseid: OTP LENS = 0xcc
<6>[    0.581878] c0      1 [CAM] ov13850_read_fuseid: OTP Sensor Version = 0xf0
<6>[    0.581889] c0      1 [CAM] ov13850_read_fuseid: OTP Driver IC Vendor & Version = 0x5d
<6>[    0.581900] c0      1 [CAM] ov13850_read_fuseid: OTP Actuator vender ID & Version = 0x4a

Adding printks to places of interest makes the problem go away, which seems
to indicate it may somehow be timing related. It also makes debugging it a lot
more fun (not really).

Work around the problem by delaying the first read of OTP data to the first
open() of the device node (i.e. startup of mm-qcamera-daemon).

Change-Id: I2ece709b61734532b848f1da9645844b40fc6bee